requirements.txt
torch
torchvision
pandas
pillow
streamlit
opencv-python 
numpy